Move id.rs into node.rs
