+++ src/flexbox.rs
@@ -3,8 +3,7 @@ use core::f32;
 use crate::flexbox::Number::{Defined, Undefined};
 use crate::forest::{Forest, NodeData};
 use crate::geometry::{Point, Rect, Size};
-use crate::id::NodeId;
-use crate::node::MeasureFunc;
+use crate::node::{MeasureFunc, NodeId};
 use crate::number::{MinMax, OrElse};
 use crate::prelude::Number;
 use crate::result;
