+++ src/lib.rs
@@ -17,7 +17,6 @@ pub mod style;
 
 mod flexbox;
 mod forest;
-mod id;
 mod sys;
 
 #[cfg(feature = "std")]
